server/ai-matching-microservice/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
websockets>=10.0,<13.0
